section_scanner/app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configparser==7.1.0
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.1.0
fix-busted-json==0.0.18
Flask==3.1.0
Flask-Cors==5.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
imageio==2.36.0
imutils==0.5.4
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
openai==1.55.0
opencv-python==4.10.0.84
packaging==24.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pywin32==308
pyzbar==0.1.9
pyzmq==26.2.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
segno==1.6.1
shapely==2.0.6
shutils==0.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tifffile==2024.9.20
tiktoken==0.8.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.23
uuid==1.30
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
